requirements.txt
phidata
yfinance
groq
duckduckgo-search
Playground